pyproject.toml
[project]
name = "obsidian-index-service"
version = "0.1.0"
description = "A service that indexes Obsidian vault notes metadata into SQLite"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "watchdog>=3.0.0",
    "python-frontmatter>=1.0.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-bdd>=6.0.0",
    "pytest-mock>=3.11.1",
    "black>=23.0.0",
    "isort>=5.12.0"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["obsidian_index_service*"]
exclude = ["tests*"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 100
